+++ pixleesdk/src/main/java/com/pixlee/pixleesdk/ui/activity/ImageViewerActivity.java
@@ -1,11 +1,9 @@
-package com.pixlee.pixleesdk.ui;
+package com.pixlee.pixleesdk.ui.activity;
 
 import android.content.Context;
 import android.content.Intent;
-import android.graphics.Color;
 import android.graphics.drawable.Drawable;
 import android.net.Uri;
-import android.os.Build;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.ImageView;
@@ -24,6 +22,7 @@ import com.pixlee.pixleesdk.PXLPhoto;
 import com.pixlee.pixleesdk.PXLPhotoSize;
 import com.pixlee.pixleesdk.PXLProduct;
 import com.pixlee.pixleesdk.databinding.ActivityImageViewerBinding;
+import com.pixlee.pixleesdk.ui.adapter.ProductAdapter;
 import com.pixlee.pixleesdk.util.PXLViewUtil;
 
 import jp.wasabeef.glide.transformations.BlurTransformation;
